+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed the issue where macOS privacy permission dialog appears repeatedly when opening articles. (#337)
 - Fixed the issue where GPU occupation is high when opening settings page. (#339)
+- Fixed the issue where some invalid feeds cause feed refresh to fail. (#341)
 
 ## [1.3.11] - 2025-12-26